Start adding some IV (current-voltage) relations
Every two port component must specify one direct current IV relation -
an equation that relates the current through the component to the voltage
across it. The derivatives of this relation with respect to the port
voltages (and optionally a dummy current variable) must also be specified.
These will be used when constructing the function representing the non-linear
system of equations, and the corresponding Jacobian matrix funtion, in
non-linear operating point analysis.

This adds DCIV relations for DCVoltageSource and Resistor.